Simplify logic with overriding callee in cases of TypedDict.get functions.
After poking around with this a bunch today I realized it would be much simplier to simply create
a context-specific Callable as opposed to attemping to hijack the rest of the typechecking.

The original implementation had problems in places, for example where a TypedDict had a List field.
A default empty list was not being coerced correctly.